bash/.bashrc: do not eval `ssh-agent`
In 7333377 (bash: eval `ssh-agent` for non-login shells, too,
2018-08-24), I moved an "eval `ssh-agent`" to my bashrc, which I
believed was necessary because a previous version of macOS disabled
running `ssh-add` (and also `ssh-agent`).

As of macOS Mojave, this is no longer the case, and so running "eval
`ssh-agent`" causes the system to accumulate ssh-agent processes as the
caller opens and closes shells.

Since macOS Mojave manages handling the ssh-agent process for us (and
tmux will forward the necessary environment variables), it's only up to
us to call ssh-add once for each shell we open.

In practice, this could probably be moved back to ~/.bash_profile, but
is unnecessary to do, since `ssh-add` will simply ignore keys that the
agent already knows about.