mutt: bind <C-r> to recall postponed message
Sometimes it is desireable to stop working on a message and pull it up
later. Instead of typing:

  :recall-message

Each time, bind <C-R> to do just that. This is a sensible default, since
we use 'r' to reply to a message, 'R' to reply-all, so <C-R> is a good
option on top of that.