mutt/.muttrc: localize 'B' macro within 'git mail'
This was introduced in 59995c4 (mutt: bind 'B' to force send a message,
2018-04-04) to easily resend messages out of a preprepared mbox, for
e.g., when preparing a patch series for submission.

This is only useful in exactly the above scenario, and has caused
significant pain and mistakes when, for e.g., resending an email from
your coworker to a large distribution list.

Move the macro into its own '~/.muttrc.git' specifically for use with
'git-mail' and make sure that 'git-mail' uses that configuration
(instead of '~/.muttrc'). This has the effect of only making this macro
available where it can be useful, and not where it can be harmful.