bin/git-reroll: expose '--list'
Sometimes it is useful to get a list of all rerolls of a given topic.
Incidentally, we compute this list in 'git_reroll__last', but then
discard it with '... | tail -r | head -1'.

Let's expose this in 'git reroll --list', and use that to implement
'git_reroll__last' by taking the output of the above through 'tail -1'.