mutt/account: use 'security find-internet-password -w'
In 217c5ec (mutt: use GMail over IMAP, not locally, 2020-04-06), I
introduced an ugly dependency on Perl (from my '~/.muttrc', no less!) to
parse out the password field of an item out of the user's login
keychain.

But, this dependency is entirely unnecessary. '/usr/bin/security's
'find-internet-password' mode supports a flag '-w' which limits the
command's output to only contain the password field.

This is exactly what we want, so use that instead of manually munging
through the output.