mutt/account/github.com: re-enable 'header_cache'
In de33743 (mutt/account/github.com: show labels, 2020-04-07), we
dropped support for 'header_cache' because it was reported [1] to
corrupt the GMail labels patch that was applied at that time.

However, this isn't true: it only "corrupts" the display when you have a
broken cache (which was the case, since I did not drop my header cache
locally at the time for 'github.com').

Since the header_cache makes it much faster to delete mail (after
reading it!) locally, let's drop the cache, re-enable it, and get the
best of both worlds.

[1]: https://www.mail-archive.com/mutt-dev@mutt.org/msg07596.html