mutt: bind 'E' to edit in the pager
In 6a8213b (mutt: move keybinds to mutt accounts, 2020-04-13), I broke
editing messages from the pager because we rebound 'e' to archive the
message.

This is frustrating in 'git mail' when I don't want to archive an
un-sent message, but rather I want to edit it and its headers. Instead
of having 'e' do different things in different contexts, rebind 'E' to
do the same job.