bin/git-who: search for matching upstream
Many projects are changing their default branch name, and so it makes
sense to try a variety of upstreams when looking for a matching author.

Likewise, this fixes a long-standing annoyance I've had with projects
that use a name other than 'master' for their default branch (e.g., some
internal projects at GitHub) for which this script used to fail, but now
works.