brew/Brewfile: install Ballast.app
Ballast is a handy utility that fixes a macOS bug where the audio
balance occasionally changes without me telling it to. I must have
installed this from the web and never added it to the Brewfile, since
I've had this on syl for as long as I can remember.

Install it by default by committing it to the Brewfile.