mutt: delete messages directly
Many "mutt+Gmail" articles on the Internet recommend that you set the
macro 'e' in the index and pager to copy either the current or tagged
messages to the '=[Gmail]/All Mail' folder.

This advice is bad: Gmail does this automatically when you delete a
message over IMAP. (Side note: the only way to *actually* delete a
message is to copy it to the '=[Gmail]/Trash' folder). Moreover, this
copy is slow, so you have to wait for your IMAP connection to copy each
message over one by one.

Deleting the message in the purest sense does not incur the slowdown,
since Gmail knows how to handle these efficiently. So, bind the 'e'
macros for the index and pager in each account to delete the message,
and receive much improved performance as a result.