alacritty: use 'screen-256color' terminal
When $TERM is set to 'xterm-256color', mutt (inside of an SSH session)
can take over the terminal's colors leaving mutt unreadable, and the
terminal's colors afterwords to be slightly different.

It's not clear if this is a bug in Alacritty or iTerm's xterm emulation.
One is definitely doing the right thing and the other is doing the wrong
thing, but I'm not sure which is which.

In either case, Alacritty's screen-256color doesn't have this problem
since it doesn't support resetting colors, so let's use that (since I
don't care about any xterm features, anyway).