mutt/.muttrc.base: properly create mbox's of multiple patches
Using <pipe-message> in conjunction with multiple tagged messages (e.g.,
tag a bunch of entries in the index, and then run ';A') does not create
a proper mbox, and 'git am' only applies the first message.

Using '<copy-message>' _does_ do the right thing by creating a bona-fide
mbox that 'git am' can read, so use that.

Helped-by: Jeff King <peff@peff.net>