mutt: bind 'L' to open the list archive view
Sometimes it is useful to recall an earlier part of a discussion that
happened on the mailing list, especially when not all parts of that
discussion are still in my inbox.

The nice thing to do would be to switch to my archive (Gmail's "All
Mail" pseudo-folder), and then something like:

  <limit>~(~i$MESSAGE_ID)

...where $MESSAGE_ID is the Message-Id header of some email that I want
to expand in the archive. This is preferred over re-downloading the
thread from the list archive since (1) the data is already present in my
message cache, and (2) it shows which messages I've already read.

Unfortunately, Mutt is slow to load my >130k email messages in the
archive, even when the header cache is enabled (and compiled
--with-tokyocabinet).

So, let's do the next best thing and re-download the messages from the
archive by figuring out the Message-ID of any email that I want to
view in context with the rest of its thread, and query the list for that
thread. Then, use my 'git lore' program to open that thread in mutt.

So, mutt is running itself as a child process, which is horrifying. But,
it works. So, let's do that.