mutt/account/github.com: appropriately set From header
When replying to messages on the Git mailing list, I want to use my
"me@ttaylorr.com" address, since that is the identity that I use when
contributing to the Git project.

This hook has been broken since it was introduced, since send2-hook
doesn't catch the outgoing mail before it's too late. The appropriate
time to do it is when composing a reply, and that is accomplished with
'reply-hook'.