alacritty: add SHA-1, SHA-256 copy hint
This super useful feature from tmux-fingers[1] has been sorely missed by
me since moving to Alacritty (and subsequently, no longer using tmux).

But it can be emulated with Alacritty's own "hints" configuration. Here,
a hint is added to look for SHA-alike output in my terminal, and copy it
when action is taken. The action can either be a mouse click, or the
"fingers"-style '^s' stroke.

[1]: https://github.com/Morantron/tmux-fingers