bin/git-apply-to: initial commit
I often want to apply a series of patches from the mailing list onto a
new topic branch (sometimes based off of an explicit upstream).

There is already good tooling for preparing an mbox of the patches I
want to apply ('D' in mutt clears any existing such mbox, 'A' applies
the current mail to an existing mbox, and ';A' applies that to all
tagged messages).

This small program applies the special mbox stored in "~/patch" to a new
branch in one step.