git: initial commit of 'git sequence'
When in the middle of some long-running operation in Git, I sometimes
want to abort whatever it is that I am doing and start over. Usually I
hit ^R and search for 'abort'. This can lead to frustrating results when
the last thing I aborted was, say, a merge, and I'm currently in the
middle of some other operation, like a rebase.

This commit introduces `git sequence`, which takes any number of
arguments and passes it to the appropriate sequencer command (rebase,
merge, cherry-pick, am, revert, etc).

That way `git sequence` (or `git seq`, as I have aliased it) will always
do the right thing when running `git sequence --abort`, no matter what
operation I'm in the middle of.

The precedence rules are a little funky, but it should be good enough to
work for most purposes. Note that this script will not run in non-Git
repositories, and it will happily accept the special `--debug' option to
print out what mode it thinks the current repository is in.