alacritty: fix typo
The 'shell' section should be at the top-level in the YAML structure,
not under 'cursor' as it has been since 0f72395 (alacritty: install on
macOS, 2020-09-17).

It doesn't seem to have caused any problems (likely because I chsh'd my
shell to /usr/local/bin/zsh anyway), but let's update it nonetheless so
that this can be installed onto a machine that doesn't store my
preferred shell.