Makefile: infer 'MACOS' from system uname
'make' by default (with no arguments) fails on macOS because this
repository assumes that it's running on Linux unless told otherwise.

But we can infer that we're on a macOS system by looking at the output
of `uname -s` on the shell, and setting MACOS if it isn't set already.
That makes `make` work out-of-the-box on macOS, which is nice.