alacritty: bind 'Cmd+N' to 'CreateNewWindow'
Alacritty version 0.10.0 ships with support for having multiple windows
managed by the same process. But the default binding for 'Cmd+N' still
causes a new Alacritty process to be spawned.

Override this behavior by telling Alacritty to explicitly run the
'CreateNewWindow' action, which adds a new window to the current
process, rather than creating a new process entirely.

(This *shouldn't* override all of the other default key_bindings
according to the documentation, but we'll see...)