bin/tailserve: initial commit
I often want a convenient way to preview, say, HTML documentation
written for the Git project. Since I don't have a browser conveniently
hooked up to my workstation, I usual resort to rsyncing the HTML file
that I'm modifying back to my laptop and opening it in my browser.

Surely there is a better way. This patch introduces a new "tailserve"
script, which starts an HTTP server that serves the current directory,
and which is bound to my Wireguard IP, so it's not open to the rest of
the internet.