mutt/.muttrc.base: propagate $TERM down to $editor
In the previous commit, aaeb29b (mutt/.muttrc.base: sort threads by
date, messages in thread-order, 2022-02-12), we should have propagated
the environment variable:

    TERM=xterm256-color

down to the $editor used by mutt, but didn't. Rectify that in this
patch, which makes it so that Vi correctly displays colors when
communicating with the terminal when invoked by Mutt.