bash/.bashrc: add 'mgit', 'mgh' aliases
I often want to read my email with a single focus depending on what time
of day it is, or what I happen to be working on.

Usually this is accomplished by running mutt, and then entering
something like:

    <limit>~ygit.git

upon startup, where `<limit>` is accomplished by hitting "l".

This works fine, but it would be convenient to just have Mutt lay filter
my messages accordingly on startup depending on how it was invoked.
After fumbling with `mutt`'s `-e` option, I realized that you can enter
functions via the usual keyboard buffer with `<push>`.

Use this to set up two new Mutt variants: one for opening my inbox
focused on the Git mailing list (with `mgit`), and another for focusing
on everything that isn't the mailing list (with `mgh`).