editorconfig: wrap EDIT_DESCRIPTION at 72 chars
The `git branch --edit-description` command opens a file
`$GIT_DIR/EDIT_DESCRIPTION` to populate the branch description, which
then becomes a cover letter.

That CL should be wrapped at 72 characters wide, so set an editorconfig
rule to wrap the EDIT_DESCRIPTION file accordingly.

(It would be nice to get rid of editorconfig entirely, since I think all
of this can and should be accomplished with my .vimrc directly, but that
is a task for another day.)