zsh/.zshrc: correctly detect when we're detached
When we initially enter a detached HEAD state, the output of "git
branch" will contain "detached at...", which we use to determine that
we're detached.

But when subsequent commits are made on top, the phrase changes from
"detached at..." to "detached from...", and parse_git_branch()
(incorrectly) concludes that we're no longer detached.

Fix this by just looking for "detached". This isn't perfect, either,
since we could have a branch that contains "detached" in it, too, but
it's good enough.