git/.gitconfig: tolerate missing diff-highlight script
The following commit is going to only conditionally build the
'diff-highlight' executable and link it into place. That will allow us
to install these dotfiles on machines that don't have a copy of git.git
sitting at $HOME/src/git.

Tolerate a missing diff-highlight executable by passing the input to the
various pagers through cat in that case.